+++ Assets/com.gamedeveducation.aibehaviours/Runtime/FrameworkEssentials/AI/Scripts/Common/BaseAI.cs
@@ -31,7 +31,6 @@ public abstract class BaseAI : MonoBehaviour
     [SerializeField] bool DEBUG_DrawLookDirection = true;
     [SerializeField] Transform DEBUG_LookTargetToSet;
     [SerializeField] bool DEBUG_UpdateLookTarget = false;
-    [field: SerializeField] public bool DEBUG_ShowSensorBounds { get; protected set; } = true;
 
     Transform CurrentLookTarget;
     Transform PreviousLookTarget;
@@ -40,68 +39,6 @@ public abstract class BaseAI : MonoBehaviour
     float DesiredPitchDelta = 0f;
     float CurrentYawDelta = 0f;
     float CurrentPitchDelta = 0f;
-
-    public Vector3 EyeLocation => HeadBoneTransform.position;
-    public Vector3 EyeDirection => HeadBoneTransform.forward;
-
-    public float VisionConeAngle => _VisionConeAngle;
-    public float VisionConeRange => _VisionConeRange;
-    public Color VisionConeColour => _VisionConeColour;
-
-    public float HearingRange => _HearingRange;
-    public Color HearingRangeColour => _HearingRangeColour;
-
-    public float ProximityDetectionRange => _ProximityDetectionRange;
-    public Color ProximityDetectionColour => _ProximityRangeColour;
-
-    public float CosVisionConeAngle { get; private set; } = 0f;
-
-    protected AwarenessSystem Awareness;
-
-    void Awake()
-    {
-        CosVisionConeAngle = Mathf.Cos(VisionConeAngle * Mathf.Deg2Rad);
-        Awareness = GetComponent<AwarenessSystem>();
-    }
-
-    public void ReportCanSee(DetectableTarget seen)
-    {
-        Awareness.ReportCanSee(seen);
-    }
-
-    public void ReportCanHear(GameObject source, Vector3 location, EHeardSoundCategory category, float intensity)
-    {
-        Awareness.ReportCanHear(source, location, category, intensity);
-    }
-
-    public void ReportInProximity(DetectableTarget target)
-    {
-        Awareness.ReportInProximity(target);
-    }
-
-    public virtual void OnSuspicious()
-    {
-    }
-
-    public virtual void OnDetected(GameObject target)
-    {
-    }
-
-    public virtual void OnFullyDetected(GameObject target)
-    {
-    }
-
-    public virtual void OnLostDetect(GameObject target)
-    {
-    }
-
-    public virtual void OnLostSuspicion()
-    {
-    }
-
-    public virtual void OnFullyLost()
-    {
-    }
 
     protected virtual void Update()
     {
@@ -173,33 +110,3 @@ public abstract class BaseAI : MonoBehaviour
         CurrentLookTarget = newLookTarget;
     }
 }
-
-#if UNITY_EDITOR
-[CustomEditor(typeof(BaseAI))]
-public class BaseAIEditor : Editor
-{
-    public void OnSceneGUI()
-    {
-        var ai = target as BaseAI;
-
-        if (ai.DEBUG_ShowSensorBounds)
-        {
-            // draw the detectopm range
-            Handles.color = ai.ProximityDetectionColour;
-            Handles.DrawSolidDisc(ai.transform.position, Vector3.up, ai.ProximityDetectionRange);
-
-            // draw the hearing range
-            Handles.color = ai.HearingRangeColour;
-            Handles.DrawSolidDisc(ai.transform.position, Vector3.up, ai.HearingRange);
-
-            // work out the start point of the vision cone
-            Vector3 startPoint = Mathf.Cos(-ai.VisionConeAngle * Mathf.Deg2Rad) * ai.transform.forward +
-                                 Mathf.Sin(-ai.VisionConeAngle * Mathf.Deg2Rad) * ai.transform.right;
-
-            // draw the vision cone
-            Handles.color = ai.VisionConeColour;
-            Handles.DrawSolidArc(ai.transform.position, Vector3.up, startPoint, ai.VisionConeAngle * 2f, ai.VisionConeRange);
-        }
-    }
-}
-#endif // UNITY_EDITOR