+++ Assets/com.gamedeveducation.aibehaviours/Runtime/FrameworkEssentials/AI/Scripts/SimpleAICharacter.cs
@@ -5,11 +5,3 @@ using UnityEditor;
 public class SimpleAICharacter : BaseAI
 {
 }
-
-
-#if UNITY_EDITOR
-[CustomEditor(typeof(SimpleAICharacter))]
-public class SimpleAICharacterEditor : BaseAIEditor
-{
-}
-#endif // UNITY_EDITOR